Kaung 12-12-22T22:18

+++ src/main/java/sg/nus/iss/team2/repository/CompensationRequestRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import sg.nus.iss.team2.model.CompensationRequest;
 
-public interface CompensationRequestRepository extends JpaRepository<CompensationRequest,Integer> {
+public interface CompensationRequestRepository extends JpaRepository<CompensationRequest,Long> {
     
 }
