+++ src/main/java/sg/nus/iss/team2/repository/LeaveBalanceRepository.java
@@ -3,6 +3,6 @@ package sg.nus.iss.team2.repository;
 import sg.nus.iss.team2.model.LeaveBalance;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaveBalanceRepository extends JpaRepository<LeaveBalance,Integer> {
+public interface LeaveBalanceRepository extends JpaRepository<LeaveBalance,Long> {
 
 }
