+++ src/main/java/sg/nus/iss/team2/service/CompensationRequestServiceImpl.java
@@ -25,7 +25,7 @@ public class CompensationRequestServiceImpl implements CompensationRequestServic
 
     @Override
     @Transactional
-    public CompensationRequest findCompensationRequest(Integer compensationRequestId){
+    public CompensationRequest findCompensationRequest(Long compensationRequestId){
         CompensationRequest cReq = crr.findById(compensationRequestId).orElse(null);
         return cReq;
     };
